Move `Game` struct and implementation out into its own file
