+++ src/entities/mod.rs
@@ -1,2 +1,3 @@
+pub mod game;
 pub mod player;
 pub mod bullet;
